Move test and swap implementation to reuse ChangeMethodName
