+++ src/test/java/org/openrewrite/java/testing/junit5/JUnit5MigrationTest.java
@@ -161,5 +161,33 @@ class JUnit5MigrationTest implements RewriteTest {
         );
     }
 
-
+    // edge case for deprecated use of assertEquals
+    // https://junit.org/junit4/javadoc/4.13/org/junit/Assert.html#assertEquals(java.lang.Object%5B%5D,%20java.lang.Object%5B%5D)
+    @Issue("https://github.com/openrewrite/rewrite-testing-frameworks/pull/384")
+    @Test
+    void assertEqualsWithArrayArgumentToAssertArrayEquals() {
+        //language=java
+        rewriteRun(
+          java(
+            """
+              import org.junit.Assert;
+              
+              class MyTest {
+                  void test() {
+                       Assert.assertEquals(new Object[1], new Object[1]);
+                  }
+              }
+              """,
+            """
+              import org.junit.jupiter.api.Assertions;
+              
+              class MyTest {
+                  void test() {
+                       Assertions.assertArrayEquals(new Object[1], new Object[1]);
+                  }
+              }
+              """
+          )
+        );
+    }
 }
